BackEnd/requirements.txt
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
tomli==2.2.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
win32_setctime==1.2.0
